Py2->3: Ensure consistent use of integer division
In python2, 3/2=1   and 3.0/2=1.5 and 4/2=2
In python3, 3/2=1.5 and 3.0/2=1.5 and 4/2=2.0
That is, python3 automatically converts operands to a float and returns a
float, whereas python2 performs integer division (aka floor division) if
the operands are integers.

Nearly all usage of division in pymapconv.py is for scaling and tiling by
evenly-divisible powers of 2, so we want integer results rather than float
results.

So, this commit changes almost all division to integer/floor division
operations.